Changed n_its from short to int; Removed print of the matrix; Increased global and local reduction sizes, but doens't seem to change anything. We are 6 times too slow right now.

+++ A4/global_vars.h
@@ -10,5 +10,5 @@
 
 /////////////////////////////////////////////
 // Global variables
-short n_its = DEFAULT_ITS;
+int n_its = DEFAULT_ITS;
 float diff_const = DEFAULT_DIFFUSION_CONSTANT;
